requirements.txt
NorenRestApiPy
requests
lxml
xmlschema
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2